requirements.txt
aiohttp
aiosmtplib
python-dotenv